+++ backend/routes/api.php
@@ -14,6 +14,7 @@ use App\Http\Controllers\Facturation\RelanceController;
 use App\Http\Controllers\Planning\MissionController;
 use App\Http\Controllers\Planning\TacheCommentaireController;
 use App\Http\Controllers\Planning\TacheController;
+use App\Http\Controllers\Portail\PortailController;
 use App\Http\Controllers\Prestations\PrestationController;
 use App\Http\Controllers\Settings\SettingController;
 use Illuminate\Support\Facades\Route;
@@ -99,7 +100,7 @@ Route::prefix('v1')->group(function () {
 
         // Portail client
         Route::middleware('portail')->prefix('portail')->group(function () {
-            //
+            Route::get('me', [PortailController::class, 'me']);
         });
     });
 });